refactor(utxorpc): spec 0.18.1

+++ txBuilding/Backend/MaestroChainContext/MaestroChainContext.go
@@ -139,10 +139,10 @@ func (mcc *MaestroChainContext) LatestEpochParams() (Base.ProtocolParameters, er
 		return protocolParams, err
 	}
 	// Map ALL the fields
-	protocolParams.MinFeeConstant = int(
+	protocolParams.MinFeeConstant = int64(
 		ppFromApi.Data.MinFeeConstant.LovelaceAmount.Lovelace,
 	)
-	protocolParams.MinFeeCoefficient = int(ppFromApi.Data.MinFeeCoefficient)
+	protocolParams.MinFeeCoefficient = int64(ppFromApi.Data.MinFeeCoefficient)
 	protocolParams.MaxTxSize = int(ppFromApi.Data.MaxTransactionSize.Bytes)
 	protocolParams.MaxBlockSize = int(ppFromApi.Data.MaxBlockBodySize.Bytes)
 	protocolParams.MaxBlockHeaderSize = int(
